src/postgres_service/requirements.txt
fastapi[all]
sqlmodel
psycopg2
python-dotenv
httpx
pytest
pydantic-extra-types
pydantic[email]
phonenumbers
sqlalchemy
sqlalchemy-utils
passlib
pyjwt